ai-app/requirements.txt
python-dotenv==1.1.1
openai==1.97.0
rich==14.0.0